task3/requirements.txt
torch>=1.9.0
transformers>=4.20.0
pandas>=1.3.0
openpyxl>=3.0.9
numpy>=1.21.0
stable-baselines3>=1.6.0
scikit-learn>=1.0.0
pyyaml>=5.4.1